#200771554 developer can add time entry to a project backend

+++ src/project/TimeEntry.kt
@@ -2,8 +2,9 @@ package project
 
 import db.Entity
 import db.Id
+import klite.Decimal
 import users.User
-import java.math.BigDecimal
+import java.math.BigInteger
 import java.time.LocalDate
 
 data class TimeEntry(
@@ -11,8 +12,8 @@ data class TimeEntry(
   val userId: Id<User>,
   val date: LocalDate = LocalDate.now(),
   val hours: Float,
-  val storyId: Int? = null,
+  val storyId: BigInteger? = null,
   val description: String? = null,
-  val hourlyRate: BigDecimal,
+  val hourlyRate: Decimal,
   override val id: Id<TimeEntry> = Id()
 ): Entity<TimeEntry>
